requirements.txt
alembic==1.13.2
fastapi==0.112.1
FastAPI-SQLAlchemy==0.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
mysql-connector-python==9.0.0
uvicorn==0.30.6
mysqlclient==2.2.4
bcrypt==4.2.0
cryptography==43.0.0
httpx==0.27.0
requests==2.32.3